combine TemplateList/DashboardTemplateList into shared component
