+++ awx/ui_next/src/components/TemplateList/TemplateListItem.test.jsx
@@ -2,12 +2,12 @@ import React from 'react';
 
 import { createMemoryHistory } from 'history';
 import { act } from 'react-dom/test-utils';
-import { mountWithContexts } from '../../../../testUtils/enzymeHelpers';
-import { JobTemplatesAPI } from '../../../api';
-import mockJobTemplateData from '../shared/data.job_template.json';
+import { mountWithContexts } from '../../../testUtils/enzymeHelpers';
+import { JobTemplatesAPI } from '../../api';
+import mockJobTemplateData from './data.job_template.json';
 import TemplateListItem from './TemplateListItem';
 
-jest.mock('../../../api');
+jest.mock('../../api');
 
 describe('<TemplateListItem />', () => {
   test('launch button shown to users with start capabilities', () => {
