+++ awx/ui_next/src/components/TemplateList/index.js
@@ -1,2 +1,2 @@
-export { default as TemplateList } from './TemplateList';
+export { default } from './TemplateList';
 export { default as TemplateListItem } from './TemplateListItem';
