+++ awx/ui_next/src/screens/Template/Templates.jsx
@@ -5,7 +5,7 @@ import { Route, withRouter, Switch } from 'react-router-dom';
 import { PageSection } from '@patternfly/react-core';
 
 import ScreenHeader from '../../components/ScreenHeader/ScreenHeader';
-import { TemplateList } from './TemplateList';
+import TemplateList from '../../components/TemplateList';
 import Template from './Template';
 import WorkflowJobTemplate from './WorkflowJobTemplate';
 import JobTemplateAdd from './JobTemplateAdd';
